+++ src/main/java/com/arangodb/internal/velocystream/VstConnectionFactorySync.java
@@ -36,7 +36,7 @@ public class VstConnectionFactorySync implements ConnectionFactory {
 
 	private final VstConnectionSync.Builder builder;
 
-	public VstConnectionFactorySync(final HostDescription host, final Integer timeout, final Long connectionTtl,
+	public VstConnectionFactorySync(final Integer timeout, final Long connectionTtl,
 		final Boolean useSsl, final SSLContext sslContext) {
 		super();
 		builder = new VstConnectionSync.Builder().timeout(timeout).ttl(connectionTtl).useSsl(useSsl)