+++ src/test/java/com/arangodb/ArangoDBTest.java
@@ -430,11 +430,6 @@ public class ArangoDBTest {
 		new ArangoDB.Builder().loadProperties(ArangoDBTest.class.getResourceAsStream("/arangodb-bad.properties"));
 	}
 
-	@Test(expected = ArangoDBException.class)
-	public void loadproperties2() {
-		new ArangoDB.Builder().loadProperties(ArangoDBTest.class.getResourceAsStream("/arangodb-bad2.properties"));
-	}
-
 	@Test
 	public void accessMultipleDatabases() {
 		String db1 = "multipledb1";